backend/requirements.txt
Django
djangorestframework
drf-yasg
redis
celery
django-celery-beat